+++ src/main/java/baseball/common/exception/InvalidLengthException.java
@@ -1,10 +1,10 @@
-package baseball.exception;
+package baseball.common.exception;
 
 public class InvalidLengthException extends Exception{
 
     private static final String MESSAGE = "자리 숫자만 가능합니다.";
 
-    public InvalidLengthException(int m) {
-        super(String.valueOf(m) + MESSAGE);
+    public InvalidLengthException(int length) {
+        super(String.valueOf(length) + MESSAGE);
     }
 }
