+++ src/main/java/baseball/common/utils/converter.java
@@ -1,4 +1,4 @@
-package baseball.utils.converter;
+package baseball.common.utils;
 
 import java.util.ArrayList;
 import java.util.List;
